Fixed types.h

+++ libc/include/types.h
@@ -5,15 +5,15 @@
 #define NULL ((void*) 0)
 #endif
 
-typedef char		i8;
-typedef int			i16;
-typedef long		i32;
-typedef long long	i64;
+typedef char	i8;
+typedef short	i16;
+typedef int		i32;
+typedef long	i64;
 
-typedef unsigned char		u8;
-typedef unsigned int		u16;
-typedef unsigned long		u32;
-typedef unsigned long long	u64;
+typedef unsigned char	u8;
+typedef unsigned short	u16;
+typedef unsigned int	u32;
+typedef unsigned long	u64;
 
 typedef u64 usize;
